Replace float sqrt+divide with fixed point rsqrt
In cdef distortion, replace floating point sqrt and divide with a fixed
point approximation of reciprocal sqrt.

The rsqrt implementation was pulled from daala, then modified to remove
the 2nd order householder iteration.

Benchmarks:
cdef_dist_wxh_8x8       time:   [93.549 ns 93.572 ns 93.597 ns]
                        change: [-6.9962% -6.7169% -6.5117%]

PSNR Y | PSNR Cb | PSNR Cr | CIEDE2000 |   SSIM | MS-SSIM | PSNR-HVS Y |
0.0556 | -0.1273 | -0.0148 |    0.0320 | 0.0581 |  0.0069 |     0.0139 |

PSNR-HVS Cb | PSNR-HVS Cr | PSNR-HVS |   VMAF | VMAF-NEG
-0.1682 |     -0.0099 |   0.0117 | 0.0111 |   0.0035
https://beta.arewecompressedyet.com/?job=master-runagain-136672c58e21803633df72dc8f62214dd2ad6b35%402021-10-31T02%3A07%3A13.188Z&job=polynomial_only_3%402021-10-31T14%3A53%3A39.769Z

+++ src/activity.rs
@@ -60,7 +60,7 @@ impl ActivityMask {
     &self, bit_depth: usize, activity_scales: &mut Box<[DistortionScale]>,
   ) {
     for (dst, &src) in activity_scales.iter_mut().zip(self.variances.iter()) {
-      *dst = ssim_boost(src as i64, src as i64, bit_depth);
+      *dst = ssim_boost(src, src, bit_depth);
     }
   }
 }
